Bump Scala version to 2.12.6 and fix provider macros
The previous provider macros, which grab the typeclass for the Props/State members of component objects at the wrapper constructor level. In Scala 2.12.4, there was a bug that allowed illegal qualifiers (https://github.com/scala/scala/pull/6359), which allowed the macro implmementations to run without failures. But in 2.12.5, this was fixed and so all the provider macros began to fail with errors for not having a stable identifier. This PR fixes the macros by creating an intermediate variable annotated with `@uncheckedStable`, which lets us access the Props/State types even though we use an unstable path.

Fixes #142