Fix types for the useCallback hook
Implement useCallback using useMemo. Since the purpose of `useCallback`
is to guarantee _refererence_ equality, the use of React's useCallback
as the underlying implementation will require a conversion back from
`js.Function` to `scala.FunctionN` and that will create a new object
every time and thus not serve the intended purpose.