Add docs on introducing React as a dependency (#238)
* Being a bit more explicit

As a newbie in Scala and Scala.js, I suggest to add more information as it is not so simple as adding the lines described above. I have tried to do it following what is explained in https://github.com/japgolly/scalajs-react/blob/master/doc/USAGE.md 

Hope you find it useful. Please, feel free to change or even ignore it, at your discretion.

Regards.
P

* Update instructions to match latest versions

* Improve clarity around when ScalaJSPlugin is used

Co-authored-by: Shadaj Laddad <shadaj@users.noreply.github.com>